+++ workbench/libs/mathtrans/spcosh.c
@@ -18,11 +18,11 @@
 
     NAME */
 
-      AROS_LH1(LONG, SPCosh,
+      AROS_LH1(float, SPCosh,
 
 /*  SYNOPSIS */
 
-      AROS_LHA(LONG, fnum1, D0),
+      AROS_LHA(float, fnum1, D0),
 
 /*  LOCATION */
